+++ src/DDTrace/Integrations/Curl/CurlIntegration.php
@@ -3,11 +3,11 @@
 namespace DDTrace\Integrations\Curl;
 
 use DDTrace\Configuration;
-use DDTrace\Formats;
+use DDTrace\Format;
 use DDTrace\Http\Urls;
 use DDTrace\Span;
-use DDTrace\Tags;
-use DDTrace\Types;
+use DDTrace\Tag;
+use DDTrace\Type;
 use DDTrace\Util\ArrayKVStore;
 use OpenTracing\GlobalTracer;
 
@@ -33,8 +33,8 @@ class CurlIntegration
             $tracer = GlobalTracer::get();
             $scope = $tracer->startActiveSpan('curl_exec');
             $span = $scope->getSpan();
-            $span->setTag(Tags\SERVICE_NAME, 'curl');
-            $span->setTag(Tags\SPAN_TYPE, Types\HTTP_CLIENT);
+            $span->setTag(Tag::SERVICE_NAME, 'curl');
+            $span->setTag(Tag::SPAN_TYPE, Type::HTTP_CLIENT);
 
             CurlIntegration::injectDistributedTracingHeaders($ch);
 
@@ -45,9 +45,9 @@ class CurlIntegration
 
             $info = curl_getinfo($ch);
             $sanitizedUrl = Urls::sanitize($info['url']);
-            $span->setTag(Tags\RESOURCE_NAME, $sanitizedUrl);
-            $span->setTag(Tags\HTTP_URL, $sanitizedUrl);
-            $span->setTag(Tags\HTTP_STATUS_CODE, $info['http_code']);
+            $span->setTag(Tag::RESOURCE_NAME, $sanitizedUrl);
+            $span->setTag(Tag::HTTP_URL, $sanitizedUrl);
+            $span->setTag(Tag::HTTP_STATUS_CODE, $info['http_code']);
 
             $scope->close();
             return $result;
@@ -61,7 +61,7 @@ class CurlIntegration
                     && is_array($value)
             ) {
                 // Storing data to be used during exec as it cannot be retrieved at then.
-                ArrayKVStore::putForResource($ch, Formats\CURL_HTTP_HEADERS, $value);
+                ArrayKVStore::putForResource($ch, Format::CURL_HTTP_HEADERS, $value);
             }
 
             return curl_setopt($ch, $option, $value);
@@ -74,7 +74,7 @@ class CurlIntegration
                     && array_key_exists(CURLOPT_HTTPHEADER, $options)
             ) {
                 // Storing data to be used during exec as it cannot be retrieved at then.
-                ArrayKVStore::putForResource($ch, Formats\CURL_HTTP_HEADERS, $options[CURLOPT_HTTPHEADER]);
+                ArrayKVStore::putForResource($ch, Format::CURL_HTTP_HEADERS, $options[CURLOPT_HTTPHEADER]);
             }
 
             return curl_setopt_array($ch, $options);
@@ -95,11 +95,11 @@ class CurlIntegration
             return;
         }
 
-        $httpHeaders = ArrayKVStore::getForResource($ch, Formats\CURL_HTTP_HEADERS, []);
+        $httpHeaders = ArrayKVStore::getForResource($ch, Format::CURL_HTTP_HEADERS, []);
         if (is_array($httpHeaders)) {
             $tracer = GlobalTracer::get();
             $context = $tracer->getActiveSpan()->getContext();
-            $tracer->inject($context, Formats\CURL_HTTP_HEADERS, $httpHeaders);
+            $tracer->inject($context, Format::CURL_HTTP_HEADERS, $httpHeaders);
 
             curl_setopt($ch, CURLOPT_HTTPHEADER, $httpHeaders);
         }