+++ src/DDTrace/Integrations/PDO/PDOIntegration.php
@@ -2,8 +2,8 @@
 
 namespace DDTrace\Integrations\PDO;
 
-use DDTrace\Tags;
-use DDTrace\Types;
+use DDTrace\Tag;
+use DDTrace\Type;
 use OpenTracing\GlobalTracer;
 
 class PDOIntegration
@@ -34,9 +34,9 @@ class PDOIntegration
             $args = func_get_args();
             $scope = GlobalTracer::get()->startActiveSpan('PDO.__construct');
             $span = $scope->getSpan();
-            $span->setTag(Tags\SPAN_TYPE, Types\SQL);
-            $span->setTag(Tags\SERVICE_NAME, 'PDO');
-            $span->setTag(Tags\RESOURCE_NAME, 'PDO.__construct');
+            $span->setTag(Tag::SPAN_TYPE, Type::SQL);
+            $span->setTag(Tag::SERVICE_NAME, 'PDO');
+            $span->setTag(Tag::RESOURCE_NAME, 'PDO.__construct');
 
             // PHP 5.4 compatible try-catch-finally
             $thrown = null;
@@ -61,9 +61,9 @@ class PDOIntegration
         dd_trace('PDO', 'exec', function ($statement) {
             $scope = GlobalTracer::get()->startActiveSpan('PDO.exec');
             $span = $scope->getSpan();
-            $span->setTag(Tags\SPAN_TYPE, Types\SQL);
-            $span->setTag(Tags\SERVICE_NAME, 'PDO');
-            $span->setTag(Tags\RESOURCE_NAME, $statement);
+            $span->setTag(Tag::SPAN_TYPE, Type::SQL);
+            $span->setTag(Tag::SERVICE_NAME, 'PDO');
+            $span->setTag(Tag::RESOURCE_NAME, $statement);
             PDOIntegration::setConnectionTags($this, $span);
 
             // PHP 5.4 compatible try-catch-finally
@@ -95,9 +95,9 @@ class PDOIntegration
             $args = func_get_args();
             $scope = GlobalTracer::get()->startActiveSpan('PDO.query');
             $span = $scope->getSpan();
-            $span->setTag(Tags\SPAN_TYPE, Types\SQL);
-            $span->setTag(Tags\SERVICE_NAME, 'PDO');
-            $span->setTag(Tags\RESOURCE_NAME, $args[0]);
+            $span->setTag(Tag::SPAN_TYPE, Type::SQL);
+            $span->setTag(Tag::SERVICE_NAME, 'PDO');
+            $span->setTag(Tag::RESOURCE_NAME, $args[0]);
             PDOIntegration::setConnectionTags($this, $span);
 
             // PHP 5.4 compatible try-catch-finally
@@ -128,8 +128,8 @@ class PDOIntegration
         dd_trace('PDO', 'commit', function () {
             $scope = GlobalTracer::get()->startActiveSpan('PDO.commit');
             $span = $scope->getSpan();
-            $span->setTag(Tags\SPAN_TYPE, Types\SQL);
-            $span->setTag(Tags\SERVICE_NAME, 'PDO');
+            $span->setTag(Tag::SPAN_TYPE, Type::SQL);
+            $span->setTag(Tag::SERVICE_NAME, 'PDO');
             PDOIntegration::setConnectionTags($this, $span);
 
             // PHP 5.4 compatible try-catch-finally
@@ -156,9 +156,9 @@ class PDOIntegration
             $args = func_get_args();
             $scope = GlobalTracer::get()->startActiveSpan('PDO.prepare');
             $span = $scope->getSpan();
-            $span->setTag(Tags\SPAN_TYPE, Types\SQL);
-            $span->setTag(Tags\SERVICE_NAME, 'PDO');
-            $span->setTag(Tags\RESOURCE_NAME, $args[0]);
+            $span->setTag(Tag::SPAN_TYPE, Type::SQL);
+            $span->setTag(Tag::SERVICE_NAME, 'PDO');
+            $span->setTag(Tag::RESOURCE_NAME, $args[0]);
             PDOIntegration::setConnectionTags($this, $span);
 
             // PHP 5.4 compatible try-catch-finally
@@ -185,9 +185,9 @@ class PDOIntegration
             $params = func_get_args();
             $scope = GlobalTracer::get()->startActiveSpan('PDOStatement.execute');
             $span = $scope->getSpan();
-            $span->setTag(Tags\SPAN_TYPE, Types\SQL);
-            $span->setTag(Tags\SERVICE_NAME, 'PDO');
-            $span->setTag(Tags\RESOURCE_NAME, $this->queryString);
+            $span->setTag(Tag::SPAN_TYPE, Type::SQL);
+            $span->setTag(Tag::SERVICE_NAME, 'PDO');
+            $span->setTag(Tag::RESOURCE_NAME, $this->queryString);
             PDOIntegration::setStatementTags($this, $span);
 
             // PHP 5.4 compatible try-catch-finally
@@ -280,10 +280,10 @@ class PDOIntegration
                     $tags['db.name'] = $value;
                     break;
                 case 'host':
-                    $tags[Tags\TARGET_HOST] = $value;
+                    $tags[Tag::TARGET_HOST] = $value;
                     break;
                 case 'port':
-                    $tags[Tags\TARGET_PORT] = $value;
+                    $tags[Tag::TARGET_PORT] = $value;
                     break;
             }
         }